tmux, vim: enable italics
Enable italics in vim as well as in tmux sessions.

For tmux use the custom TERMINFO 'tmux-256color' for $TERM (see
'/tmux/tmux.terminfo' which needs to be installed to the machine using
`tic` for this to function properly).

For vim (outside of tmux) the basic 'xterm-256color' TERMINFO doesn't
properly handle italics so we just explicitly tell vim which escape
codes to use.